+++ CHANGELOG.md
@@ -11,3 +11,4 @@ See [keep a changelog](https://keepachangelog.com/en/1.0.0/) for information abo
 * Added Sprint Report.
 * Added Planning.
 * Added OIDC login.
+* Added Project Billing.